# scratch/extracted_replit/Hello-World/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Libyan E-Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        fullName:
          type: string
        fullNameAr:
          type: string
        phoneNumber:
          type: string
        role:
          type: string
          enum: [student, teacher]
        language:
          type: string
          enum: [ar, en]
          default: ar
        passkey:
          type: string
          nullable: true
        tier:
          type: string
          nullable: true
        agreedToCommission:
          type: boolean
          nullable: true
      required:
        - email
        - password
        - fullName
        - role
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        fullNameAr:
          type: string
        role:
          type: string
          enum: [student, teacher, admin]
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        bioAr:
          type: string
          nullable: true
        language:
          type: string
          enum: [ar, en]
        hasPasskey:
          type: boolean
        biometricsVerified:
          type: boolean
        phoneVerified:
          type: boolean
        emailVerified:
          type: boolean
        balance:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - fullName
        - role
        - language
        - createdAt
...